+++ react/package.json
