+++ react/src/vite-env.d.ts
