+++ react/vite.config.ts
@@ -1,0 +1,7 @@
+import react from "@vitejs/plugin-react-swc";
+import { defineConfig } from "vite";
+
+// https://vitejs.dev/config/
+export default defineConfig({
+	plugins: [react()],
+});
